# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Reimbursement Prep API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/search
        - $ref: ./openapi_components_core.yaml#/components/parameters/archived
        - $ref: ./openapi_components_core.yaml#/components/parameters/pinned
        - $ref: ./openapi_components_core.yaml#/components/parameters/cursor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/ProjectList
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ProjectCreateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Project
  /projects/{project_id}:
    patch:
      summary: Update project
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ProjectUpdateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Project
  /projects/{project_id}/archive:
    post:
      summary: Archive project
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
      responses:
        "200":
          description: OK
  /projects/{project_id}/expenses:
    get:
      summary: List expenses
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
        - $ref: ./openapi_components_core.yaml#/components/parameters/status
        - $ref: ./openapi_components_core.yaml#/components/parameters/category
        - $ref: ./openapi_components_core.yaml#/components/parameters/date_from
        - $ref: ./openapi_components_core.yaml#/components/parameters/date_to
        - $ref: ./openapi_components_core.yaml#/components/parameters/cursor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/ExpenseList
    post:
      summary: Create expense
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/project_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ExpenseCreateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Expense
  /expenses/{expense_id}:
    patch:
      summary: Update expense
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/expense_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./openapi_components_requests.yaml#/components/schemas/ExpenseUpdateRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/Expense
  /expenses/{expense_id}/receipt-candidates:
    get:
      summary: Get receipt match candidates
      parameters:
        - $ref: ./openapi_components_core.yaml#/components/parameters/expense_id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: ./openapi_components_core.yaml#/components/schemas/CandidateList
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT